+++ dev/ts/dev-bundle-index.ts
@@ -14,6 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export * from './utils/main';
-export * from './utils/shared-helpers';
+export * from './shared/main';
+export * from './shared/shared-helpers';
 export * from '../../src/bpmn-visualization';
