+++ dev/ts/shared/internal-helpers.ts
