+++ dev/ts/shared/shared-helpers.ts
